+++ clang/include/clang/Sema/ParseHLSLRootSignature.h
